Initial commit
The print function works (and looks pretty awesome), but either the
read-in or next board function is messing up (every other board contains
about 8 characters from the previous board for some reason). I will be
working on fixing that issue this evening (probably). Its just a start,
but it should have most of the stuff we will need to do already defined,
just not filled out yet.