file reading works now
Completely redid file read in function. Instead of getting each line, it
now reads in individual characters and keeps track of where it is in the
board matrix. Also, defaults to read in sudoku.txt if there is an issue
with the file the user specified to open. Also, added a bunch of
comments.